requirements.txt
Django>=4.2,<6
djangorestframework>=3.14,<4
psycopg2-binary>=2.9,<3
dj-database-url>=2.1,<4
langchain>=0.1.20,<0.2
huggingface-hub>=0.20,<1
scikit-learn>=1.3,<2
python-dotenv>=1.0,<2
gunicorn>=21,<24
django-jazzmin>=3.0,<4
